requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# File parsing
pdfplumber==0.10.3
pypdf==3.17.4
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2

# Calendar generation
icalendar==5.0.11
pytz==2024.1

# OCR (optional, for image-based PDFs)
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS support
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
